# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todolist Sorter",
    "description": "Self-hosted service that watches a Todoist project for new items and reorders them into a user-defined category sequence using an LLM. Optionally also transforms item content (typo fixes, emoji) based on per-project `additional_instructions`.\n\n**Auth.** Management endpoints require an `X-API-Key: <APP_API_KEY>` header. The webhook, OAuth callback, health probe, and `/setup` pages are public.\n\nFirst-time setup: open `/setup` in a browser.",
    "contact": {
      "name": "Source on GitHub",
      "url": "https://github.com/mcdax/todolist-sorter"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/oauth/callback": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Todoist OAuth redirect target",
        "description": "Redirect target for the Todoist OAuth authorise flow. Exchanges the `code` query parameter for an access token so Todoist marks the app as \"installed\" and starts delivering webhooks. The access token itself is discarded; the service uses the personal `TODOIST_API_TOKEN` from the environment for all subsequent Todoist API calls. On success, a marker file is written so `/setup` can display the authorised state.",
        "operationId": "callback_oauth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth flow completed successfully.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Missing `code` or user denied access."
          },
          "500": {
            "description": "Todoist token exchange failed."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}